Renames, and beginning of implementing edit a time distance exercise
